refactor: share backgroundTrayHasItems between GridView and BackgroundTray
BackgroundTray already returns null when empty, and GridView duplicated
that predicate to decide whether to render the padding wrapper. Extract
a single helper both call sites use so adding a new background item
type only requires updating one place.
